+++ lib/request_handler/body_parser.rb
@@ -31,7 +31,9 @@ module RequestHandler
       b = b.read
       b.empty? ? {} : MultiJson.load(b)
     rescue MultiJson::ParseError => e
-      raise ParseError, json: e.message
+      raise BodyFormatError, [{ status: '400',
+                                code: 'INVALID_BODY_FORMAT',
+                                detail: e.message }]
     end
 
     attr_reader :request, :schema, :schema_options, :type